+++ src/PrimitiveModelLoader.hpp
@@ -2,6 +2,7 @@
 
 #include <cstdint>
 #include <vector>
+
 #include "glad/glad.h"
 #include "GLFW/glfw3.h"
 #include "glm/vec3.hpp"
@@ -10,13 +11,13 @@
 class PrimitiveModelLoader
 {
 private:
-	static unsigned int currentVAOId;
-	static unsigned int currentVBOId;
+	static uint32_t currentVAOId;
+	static uint32_t currentVBOId;
 
 	static void CreateAndBindVAO();
-	static void AddData(unsigned int attributeId, std::vector<float> data);
-	static unsigned int AddIndices(std::vector<unsigned int> indices);
+	static void AddData(uint8_t attributeId, std::vector<float> data);
+	static uint32_t AddIndices(std::vector<uint32_t> indices);
 	static void UnbindVAO();
 public:
-	static Model* CreateModel(std::vector<glm::vec3> vertices, std::vector<unsigned int> indices);
+	static Model* CreateModel(std::vector<glm::vec3> vertices, std::vector<uint32_t> indices);
 };
